requirements.txt
--extra-index-url https://download.pytorch.org/whl/cpu
transformers
torch
librosa
customtkinter
soundfile
numpy
pyinstaller
pynput
sounddevice
pyperclip
pyautogui
pystray
Pillow
